combine components and containers
