+++ react-redux-boilerplate/README.md
@@ -20,14 +20,15 @@ This project has 4 subdirectories:
 * `actions/`: **Defining state transitions**: Contains all the possible actions that can occur in an
     application. Actions just plain Javascript objects that contain the `type` of the action, and
     any metadata the `store` needs to respond to that action.
-* `components/`: **Visual display**. React `Component`s that are generally **presentational** in
-    nature—that is, they do not specify how to manage state, but merely describe what HTML should be
-    outputted given their input properties.
-    * See the README in `components/` for more detail on handling individual (not application-wide)
-        component state.
-* `containers/`: **Transferring application state to components**. Containers are wrappers around
-    components (in the `components/` directory) that pass them state and functions that modify
-    state, so that the `components` don't need to know anything about how application state is
-    managed. That way, the components can focus entirely on presentation.
+* `components/`: Each component contains two parts:
+    * Components that define **visual display**. React `Component`s that are generally
+        **presentational** in nature—that is, they do not specify how to manage state, but merely
+        describe what HTML should be outputted given their input properties.
+        * See the README in `components/` for more detail on handling individual (not
+            application-wide) component state.
+    * Containers that **transfer application state to components**. Containers are wrappers
+        around presentational components that pass them state and functions that
+        modify state, so that the `components` don't need to know anything about how application
+        state is managed. That way, the components can focus entirely on presentation.
 
 Enter each subdirectory and read their documentation for more information on how they work.
