+++ react-redux-boilerplate/components/CounterSection.jsx
@@ -1,10 +1,11 @@
+// Component
 // View of the section of the page that displays a counter and a button to increment it.
 // Presentational code only; state is passed as properties by the container.
 import { Button } from 'clever-components';
 import React, { PropTypes } from 'react';
 
 require('./CounterSection.less');
-export default function CounterSectionView({ counterValue, incrementCounter }) {
+export function CounterSectionView({ counterValue, incrementCounter }) {
   return (<div className="section CounterSection">
     <h2>Counter</h2>
     <p>The current value is: <span className="CounterSection--value">{counterValue}</span></p>
@@ -16,3 +17,25 @@ CounterSectionView.propTypes = {
   counterValue: PropTypes.number.isRequired,
   incrementCounter: PropTypes.func.isRequired,
 };
+
+// Container
+// Injects state and action dispatchers into the Component, thus decoupling the
+// presentation from state management.
+import { connect } from 'react-redux';
+
+import { incrementCounter } from '../actions';
+
+function mapStateToProps(state) {
+  return { counterValue: state.counter.value };
+}
+
+function mapDispatchToProps(dispatch) {
+  return { incrementCounter: () => dispatch(incrementCounter()) };
+}
+
+const CounterSection = connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(CounterSectionView);
+
+export default CounterSection;
